+++ src/components/footer.js
@@ -1,4 +1,3 @@
-import { h } from 'preact';
 import styled from 'styled-components';
 import { timeAgo } from '../utils';