+++ webpack.config.js
@@ -23,7 +23,7 @@ const config = {
         exclude: /node_modules/,
         options: {
           presets: ['env'],
-          plugins: [['transform-react-jsx', { pragma: 'h' }]]
+          plugins: ['transform-react-jsx']
         }
       },
       {
@@ -41,17 +41,20 @@ const config = {
     historyApiFallback: true,
     port: 9002
   },
-  resolve: {
-    alias: {
-      react: 'preact-compat',
-      'react-dom': 'preact-compat'
-    }
-  },
-  plugins: [new BundleAnalyzerPlugin()]
+  plugins: [
+    new webpack.ProvidePlugin({
+      React: 'react'
+    }),
+    new BundleAnalyzerPlugin()
+  ]
 };
 
 if (process.env.NODE_ENV === 'production') {
   config.plugins = [
+    new webpack.DefinePlugin({ 'process.env.NODE_ENV': JSON.stringify('production') }),
+    new webpack.ProvidePlugin({
+      React: 'react'
+    }),
     new webpack.optimize.UglifyJsPlugin({
       compress: {
         warnings: false,