+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 # Changelog
 
 ## [Unreleased][]
+### Changed
+* swapped Preact for React!
+* upgraded dependencies
 
 ### Added
 * Project setup linting/prettier/changelog